# Conflicts:
#	src/test/java/RestapiTest.java

+++ src/main/java/io/swagger/model/Frontend.java
@@ -27,7 +27,7 @@ public class Frontend implements Serializable{
   private String mode = null;
 
   @JsonProperty("bindPort")
-  private Long bindPort = null;
+  private Integer bindPort = null;
 
   @JsonProperty("bindIp")
   private String bindIp = null;
@@ -94,7 +94,7 @@ public class Frontend implements Serializable{
     this.mode = mode;
   }
 
-  public Frontend bindPort(Long bindPort) {
+  public Frontend bindPort(Integer bindPort) {
     this.bindPort = bindPort;
     return this;
   }
@@ -106,11 +106,11 @@ public class Frontend implements Serializable{
   @ApiModelProperty(value = "")
 
 
-  public Long getBindPort() {
+  public Integer getBindPort() {
     return bindPort;
   }
 
-  public void setBindPort(Long bindPort) {
+  public void setBindPort(Integer bindPort) {
     this.bindPort = bindPort;
   }
